Rename utils to ledger_utils, move arg parsing funcs to a new module
